// types/artist.ts
export type Artist = {
	id: string;
	name: string;
	avatar_url?: string | null;
	cover_url?: string | null;
	bio?: string | null;
	verified?: boolean | null;
	instagram?: string | null;
	youtube?: string | null;
	tiktok?: string | null;
	spotify?: string | null;
	apple_music?: string | null;
	x_twitter?: string | null;
};